Calculating best safe play and burn
No longer choosing the "first" in the hand. The player will now attempt
to play the card that it thinks has the lowest number value if there is
more than one to choose from.

Also added functionality to cardFact to estimate the number that is
expected from the perspective of a current player. Used when calculating
the bestSafePlay and Yolo